Improved project structure (moved files and directories around)
The main entry point is easier to find now.
